git: show-index
